requirements.txt
datasets
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
requests
python-jose[cryptography]
passlib[bcrypt] 